Inlines Cassandra autoconfiguration
