+++ zipkin-server/pom.xml
@@ -120,10 +120,10 @@
       <version>1.11.3</version>
     </dependency>
 
-    <!-- Cassandra backend -->
+    <!-- Cassandra and Cassandra 3 backends -->
     <dependency>
-      <groupId>io.zipkin.java</groupId>
-      <artifactId>zipkin-autoconfigure-storage-cassandra</artifactId>
+      <groupId>io.zipkin.zipkin2</groupId>
+      <artifactId>zipkin-storage-cassandra-v1</artifactId>
       <version>${project.version}</version>
       <optional>true</optional>
       <exclusions>
@@ -156,24 +156,6 @@
         <exclusion>
           <groupId>com.datastax.cassandra</groupId>
           <artifactId>cassandra-driver-core</artifactId>
-        </exclusion>
-      </exclusions>
-    </dependency>
-
-    <!-- Cassandra 3 backend -->
-    <dependency>
-      <groupId>io.zipkin.java</groupId>
-      <artifactId>zipkin-autoconfigure-storage-cassandra3</artifactId>
-      <version>${project.version}</version>
-      <optional>true</optional>
-      <exclusions>
-        <exclusion>
-          <groupId>com.datastax.cassandra</groupId>
-          <artifactId>cassandra-driver-core</artifactId>
-        </exclusion>
-        <exclusion>
-          <groupId>com.datastax.cassandra</groupId>
-          <artifactId>cassandra-driver-mapping</artifactId>
         </exclusion>
       </exclusions>
     </dependency>
@@ -248,6 +230,7 @@
     <dependency>
       <groupId>org.apache.zipkin.brave.cassandra</groupId>
       <artifactId>brave-instrumentation-cassandra-driver</artifactId>
+      <version>0.10.2</version>
       <optional>true</optional>
     </dependency>
