+++ zipkin-server/src/main/java/zipkin2/server/internal/InternalZipkinConfiguration.java
@@ -16,6 +16,8 @@ package zipkin2.server.internal;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import zipkin2.server.internal.brave.TracingConfiguration;
+import zipkin2.server.internal.cassandra.ZipkinCassandraStorageConfiguration;
+import zipkin2.server.internal.cassandra3.ZipkinCassandra3StorageConfiguration;
 import zipkin2.server.internal.kafka.ZipkinKafkaCollectorConfiguration;
 import zipkin2.server.internal.mysql.ZipkinMySQLStorageConfiguration;
 import zipkin2.server.internal.prometheus.ZipkinPrometheusMetricsConfiguration;
@@ -26,6 +28,8 @@ import zipkin2.server.internal.ui.ZipkinUiConfiguration;
 @Import({
   ZipkinServerConfiguration.class,
   ZipkinUiConfiguration.class,
+  ZipkinCassandraStorageConfiguration.class,
+  ZipkinCassandra3StorageConfiguration.class,
   ZipkinMySQLStorageConfiguration.class,
   TracingConfiguration.class,
   ZipkinQueryApiV2.class,
