+++ zipkin-server/src/main/java/zipkin2/server/internal/cassandra3/ZipkinCassandra3StorageConfiguration.java
@@ -11,20 +11,22 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package zipkin2.autoconfigure.storage.cassandra;
+package zipkin2.server.internal.cassandra3;
 
+import brave.Tracing;
+import brave.cassandra.driver.TracingSession;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import zipkin2.server.internal.ConditionalOnSelfTracing;
 import zipkin2.storage.StorageComponent;
-import zipkin2.storage.cassandra.v1.CassandraStorage;
-import zipkin2.storage.cassandra.v1.SessionFactory;
+import zipkin2.storage.cassandra.CassandraStorage.SessionFactory;
 
 /**
  * This storage accepts Cassandra logs in a specified category. Each log entry is expected to
@@ -32,32 +34,53 @@ import zipkin2.storage.cassandra.v1.SessionFactory;
  * are stored asynchronously.
  */
 @Configuration
-@EnableConfigurationProperties(ZipkinCassandraStorageProperties.class)
-@ConditionalOnProperty(name = "zipkin.storage.type", havingValue = "cassandra")
+@EnableConfigurationProperties(ZipkinCassandra3StorageProperties.class)
+@ConditionalOnProperty(name = "zipkin.storage.type", havingValue = "cassandra3")
 @ConditionalOnMissingBean(StorageComponent.class)
-class ZipkinCassandraStorageAutoConfiguration {
+// This component is named .*Cassandra3.* even though the package already says cassandra3 because
+// Spring Boot configuration endpoints only printout the simple name of the class
+public class ZipkinCassandra3StorageConfiguration {
 
-  @Autowired(required = false)
-  @Qualifier("tracingSessionFactory")
-  SessionFactory tracingSessionFactory;
+  @Bean SessionFactory sessionFactory() {
+    return SessionFactory.DEFAULT;
+  }
 
   @Bean
   @ConditionalOnMissingBean
   StorageComponent storage(
-      ZipkinCassandraStorageProperties properties,
+      ZipkinCassandra3StorageProperties properties,
+      SessionFactory sessionFactory,
       @Value("${zipkin.storage.strict-trace-id:true}") boolean strictTraceId,
       @Value("${zipkin.storage.search-enabled:true}") boolean searchEnabled,
       @Value("${zipkin.storage.autocomplete-keys:}") List<String> autocompleteKeys,
       @Value("${zipkin.storage.autocomplete-ttl:3600000}") int autocompleteTtl,
       @Value("${zipkin.storage.autocomplete-cardinality:20000}") int autocompleteCardinality) {
-    CassandraStorage.Builder builder = properties.toBuilder()
+    return properties.toBuilder()
       .strictTraceId(strictTraceId)
       .searchEnabled(searchEnabled)
       .autocompleteKeys(autocompleteKeys)
       .autocompleteTtl(autocompleteTtl)
-      .autocompleteCardinality(autocompleteCardinality);
-    return tracingSessionFactory == null
-        ? builder.build()
-        : builder.sessionFactory(tracingSessionFactory).build();
+      .autocompleteCardinality(autocompleteCardinality)
+      .sessionFactory(sessionFactory).build();
+  }
+
+  @Configuration
+  @ConditionalOnSelfTracing
+  static class TracingSessionFactoryEnhancer implements BeanPostProcessor {
+
+    @Autowired(required = false) Tracing tracing;
+
+    @Override public Object postProcessBeforeInitialization(Object bean, String beanName) {
+      return bean;
+    }
+
+    @Override public Object postProcessAfterInitialization(Object bean, String beanName) {
+      if (tracing == null) return bean;
+      if (bean instanceof SessionFactory) {
+        SessionFactory delegate = (SessionFactory) bean;
+        return (SessionFactory) storage -> TracingSession.create(tracing, delegate.create(storage));
+      }
+      return bean;
+    }
   }
 }
